MATH AND COMPARISON OPERATORS SUBMISSION ASSIGNMENT
C# AND UNITY COURSE Step: 130
REQUIREMENTS:

The program must start by printing “Anonymous Income Comparison Program” to the screen.

It must then print “Person 1” to the screen and get the following details:

Hourly Rate

Hours worked per week

It must then print “Person 2” to the screen and then get the following details:

Hourly rate

Hours worked per week

It must then print to the screen “Annual salary of Person 1:” and display the exact salary below it.

It must then print to the screen “Annual salary of Person 2:” and display the exact salary below it.

It must then print to the screen “Does Person 1 make more money than Person 2?” and write the true or false value of this statement below it.