CLASS SUBMISSION ASSIGNMENT
C# AND UNITY COURSE
Step: 224
Create a class. In that class, create a void method that outputs an integer. Have the method divide the data passed to it by 2.

In the Main() method, instantiate that class.

Have the user enter a number. Call the method on that number. Display the output to the screen. It should be the entered number, divided by two.

Create a method with output parameters.

Overload a method.

Declare a class to be static.