PARAMETERS SUBMISSION ASSIGNMENT
C# AND UNITY COURSE Step: 256Make the Employee class take a generic type parameter.

Add a property to the Employee class called “things” and have its data type be a generic list matching the generic type of the class.

Instantiate an Employee object with type “string” as its generic parameter. Assign a list of strings as the property value of Things.

Instantiate an Employee object with type “int” as its generic parameter. Assign a list of integers as the property value of Things.

Create a loop that prints all of the Things to the Console.