LAMBDA SUBMISSION ASSIGNMENT
C# AND UNITY COURSE Step: 273
In the Main() method, create a list of at least 10 employees. Each employee should have a first and last name, as well as an Id. At least two employees should have the first name “Joe”.

Using a foreach loop, create a new list of all employees with the first name “Joe”.

Perform the same action again, but this time with a lambda expression.

Using a lambda expression, make a list of all employees with an Id number greater than 5.